environment-discovery-20250908-070923/config/files/opt/rpa-system/requirements.txt
fastapi
uvicorn
pydantic
requests
apscheduler
sqlalchemy
tenacity
python-dotenv
bcrypt
pyjwt[crypto]
passlib[bcrypt]
cryptography
selenium
pyotp
psutil
pillow
python-multipart
aiofiles
lxml
